+++ Radia.Tests/ViewModels/RadiaFileInfoViewModelTests.cs
@@ -7,6 +7,7 @@ using Radia.Services.FileProviders.Local;
 using Radia.ViewModels;
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -24,7 +25,7 @@ namespace Radia.Tests.ViewModels
             public void WhenGivenAllDependencies_ThenAllPropertiesAreSet()
             {
                 string content = "FancyTestSite";
-                string testFile = Path.Combine(TestContext!.TestRunDirectory!, Guid.NewGuid().ToString().ToLowerInvariant().Replace('-', 'd'));
+                string testFile = Path.Combine(TestContext!.TestRunResultsDirectory!, Guid.NewGuid().ToString().ToLowerInvariant().Replace('-', 'd'));
 
                 File.WriteAllText(testFile, content);
 
@@ -49,7 +50,7 @@ namespace Radia.Tests.ViewModels
             public void WhenGivenAllDependenciesAndEmptyFile_ThenAllPropertiesAreSet()
             {
                 string content = "";
-                string testFile = Path.Combine(TestContext!.TestRunDirectory!, Guid.NewGuid().ToString().ToLowerInvariant().Replace('-', 'd'));
+                string testFile = Path.Combine(TestContext!.TestRunResultsDirectory!, Guid.NewGuid().ToString().ToLowerInvariant().Replace('-', 'd'));
 
                 File.WriteAllText(testFile, content);
 
@@ -68,31 +69,6 @@ namespace Radia.Tests.ViewModels
                 radiaFileInfoViewModel.LastModified.Should().BeCloseTo(DateTime.Now, TimeSpan.FromSeconds(10));
                 radiaFileInfoViewModel.Length.Should().Be(0);
                 radiaFileInfoViewModel.FormattedLength.Should().Be("0 bytes");
-            }
-
-            [TestMethod]
-            public void WhenGivenAllDependenciesAndDirectory_ThenAllPropertiesAreSet()
-            {
-                string content = "FancyTestSite";
-                string directoryName = "NiceDirectory\\";
-                string testDirectory = Path.Combine(TestContext!.TestRunDirectory!, directoryName);
-
-                Directory.CreateDirectory(testDirectory);
-
-                IByteSizeService byteSizeService = new ByteSizeService();
-
-                IRadiaFileProvider radiaFileProvider = new LocalFileProvider(TestContext!.TestRunDirectory!, true);
-                IRadiaFileInfo radiaFileInfo = radiaFileProvider.GetFileInfo(testDirectory);
-
-                RadiaFileInfoViewModel radiaFileInfoViewModel = new RadiaFileInfoViewModel(byteSizeService,
-                                                                                           DefaultRadiaTestContext.WebHost,
-                                                                                           radiaFileInfo,
-                                                                                           true,
-                                                                                           $"/{testDirectory.Replace("\\","/")}");
-
-                radiaFileInfoViewModel.Name.Should().Be(testDirectory);
-                radiaFileInfoViewModel.Length.Should().Be(-1);
-                radiaFileInfoViewModel.FormattedLength.Should().Be("-");
             }
         }
     }